+++ examples/qt3d/solarsystem/solarsystem.h
@@ -53,34 +53,34 @@ QT_END_NAMESPACE
 class SolarSystemView : public QGLView
 {
     Q_OBJECT
-    Q_PROPERTY(qreal angle1 READ angle1 WRITE setAngle1)
-    Q_PROPERTY(qreal angle2 READ angle2 WRITE setAngle2)
-    Q_PROPERTY(qreal angle3 READ angle3 WRITE setAngle3)
-    Q_PROPERTY(qreal glowFactor READ glowFactor WRITE setGlowFactor)
+    Q_PROPERTY(float angle1 READ angle1 WRITE setAngle1)
+    Q_PROPERTY(float angle2 READ angle2 WRITE setAngle2)
+    Q_PROPERTY(float angle3 READ angle3 WRITE setAngle3)
+    Q_PROPERTY(float glowFactor READ glowFactor WRITE setGlowFactor)
 public:
     SolarSystemView(QWindow *parent = 0);
     ~SolarSystemView();
 
-    qreal angle1() const { return m_angle1; }
-    void setAngle1(qreal angle) { m_angle1 = angle; update(); }
+    float angle1() const { return m_angle1; }
+    void setAngle1(float angle) { m_angle1 = angle; update(); }
 
-    qreal angle2() const { return m_angle2; }
-    void setAngle2(qreal angle) { m_angle2 = angle; update(); }
+    float angle2() const { return m_angle2; }
+    void setAngle2(float angle) { m_angle2 = angle; update(); }
 
-    qreal angle3() const { return m_angle3; }
-    void setAngle3(qreal angle) { m_angle3 = angle; update(); }
+    float angle3() const { return m_angle3; }
+    void setAngle3(float angle) { m_angle3 = angle; update(); }
 
-    qreal glowFactor() const { return m_glowFactor; }
-    void setGlowFactor(qreal arg) { m_glowFactor = arg; update(); }
+    float glowFactor() const { return m_glowFactor; }
+    void setGlowFactor(float arg) { m_glowFactor = arg; update(); }
 
 protected:
     void initializeGL(QGLPainter *painter);
     void paintGL(QGLPainter *painter);
 
 private:
-    qreal m_angle1;
-    qreal m_angle2;
-    qreal m_angle3;
+    float m_angle1;
+    float m_angle2;
+    float m_angle3;
     float m_glowFactor;
     QGraphicsRotation3D *sunRotation;
     QGraphicsRotation3D *planetRotation;
